requirements.txt
flet
aiohttp
sqlalchemy
python-dotenv
pymysql